#17 added middle ware to users

+++ jawbs-server/server/controllers/users.ts
@@ -1,11 +1,11 @@
-import { Request, Response } from "express";
+import { NextFunction, Request, Response } from "express";
 import prisma from "../libs/prisma.js";
+import { ClientError } from "../middlewares";
 
-const create = async (req: Request, res: Response) => {
+const create = async (req: Request, res: Response, next: NextFunction) => {
   const { firstName, lastName, email, phone, summary } = req.body;
   if (!firstName || !lastName || !email || !phone) {
-    res.status(400).json("Improper format");
-    return;
+    throw new ClientError(400, "Improper format, please submit again");
   }
   try {
     const newUser = await prisma.user.create({
@@ -23,31 +23,48 @@ const create = async (req: Request, res: Response) => {
     });
     res.status(201).json(newUser.id);
   } catch (err) {
-    res.status(400).json({ message: err });
+    next(err);
   } finally {
     await prisma.$disconnect();
   }
 };
 
-const read = async (req: Request, res: Response) => {
+const read = async (req: Request, res: Response, next: NextFunction) => {
+  const { id } = req.params;
+  if (!id) {
+    throw new ClientError(400, "Id is needed");
+  }
+
   try {
-    const { id } = req.params;
     const user = await prisma.user.findFirst({
       where: {
         id,
       },
     });
+
+    if (!user) {
+      throw new ClientError(404, "User not found");
+    }
     res.status(200).json(user);
   } catch (err) {
-    res.status(400).json({ message: err });
+    next(err);
   } finally {
     await prisma.$disconnect();
   }
 };
 
-const readPortfolio = async (req: Request, res: Response) => {
+const readPortfolio = async (
+  req: Request,
+  res: Response,
+  next: NextFunction,
+) => {
+  const { id } = req.params;
+
+  if (!id) {
+    throw new ClientError(400, "Improper format");
+  }
+
   try {
-    const { id } = req.params;
     const user = await prisma.user.findFirst({
       where: {
         id,
@@ -72,53 +89,70 @@ const readPortfolio = async (req: Request, res: Response) => {
       },
     });
     if (!user) {
-      res.status(404).json("User not found");
-      return;
+      throw new ClientError(404, "User not found");
     }
     res.status(200).json(user);
   } catch (err) {
-    res.status(400).json({ message: err });
+    next(err);
   } finally {
     await prisma.$disconnect();
   }
 };
 
-const readAll = async (__: Request, res: Response) => {
+const readAll = async (__: Request, res: Response, next: NextFunction) => {
   try {
     const user = await prisma.user.findMany();
+    if (!user) {
+      throw new ClientError(404, "No users found");
+    }
     res.status(200).json(user);
   } catch (err) {
-    res.status(400).json({ message: err });
+    next(err);
   } finally {
     await prisma.$disconnect();
   }
 };
 
-const update = async (req: Request, res: Response) => {
+const update = async (req: Request, res: Response, next: NextFunction) => {
+  const { id } = req.params;
+
+  if (!id) {
+    throw new ClientError(400, "Improper format");
+  }
   try {
-    const { id } = req.params;
     const updates = await req.body;
     const updatedUser = await prisma.user.update({
       where: { id },
       data: { ...updates },
     });
+
+    if (!updatedUser) {
+      throw new ClientError(404, "User not found");
+    }
+
     res.status(200).json(updatedUser);
   } catch (err) {
-    res.status(400).json({ message: err });
+    next(err);
   } finally {
     await prisma.$disconnect();
   }
 };
 
-const destroy = async (req: Request, res: Response) => {
+const destroy = async (req: Request, res: Response, next: NextFunction) => {
+  const { id } = req.params;
+  if (!id) {
+    throw new ClientError(400, "Improper format");
+  }
   try {
-    const { id } = req.params;
     const deletedUser = await prisma.user.delete({
       where: { id },
     });
+    if (!deletedUser) {
+      throw new ClientError(404, "User not found");
+    }
     res.status(200).json(deletedUser);
   } catch (err) {
-    res.status(400).json({ message: err });
+    next(err);
   } finally {
     await prisma.$disconnect();
   }
